Core kernel Yaiwes/Componentes recuperados A/OpenThoughts/open_thoughts_3/requirements.txt
setuptools
black
bs4
pandas
tqdm
python-box
jsonpickle
openai
transformers
fire
tiktoken
huggingface_hub[cli]
ray[default]==2.36.0
fasttext-wheel
datasets
boto3
asttokens
aiohttp[speedups]>=3.8
backoff>=2.2
requests>=2.28
docker-pycreds
sentry-sdk
loguru>=0.7
deepspeed
sentence-transformers>=2.2.2
fastapi>=0.101.0
uvicorn>=0.23.0
optimum==1.12.0
google-auth==2.25.1
fsspec==2024.9.0
pyopenssl>=24.0.0 # For some reason we get a strange error without this, see https://github.com/conda/conda/issues/13619#issuecomment-1966164903
./dcft/external_repositories/code2flow
psycopg[binary]
sagemaker
sqlalchemy
python-dotenv
shortuuid
fuzzywuzzy
codebleu
tree-sitter-python
tree-sitter-java
tenacity
jsonlines
fschat
dashscope
prettytable
scikit-learn
hf-transfer
httpx
nltk
sentencepiece
protobuf
google-cloud-aiplatform
accelerate[sagemaker]
botocore
jmespath
sacrebleu
portalocker
tabulate
colorama
lxml
evaluate
scipy
joblib
threadpoolctl
tree_sitter
termcolor
sqlitedict
psycopg2-binary
rapidfuzz
gcsfs
bespokelabs-curator@git+https://github.com/bespokelabsai/curator.git@fix_casting_ant
faiss-cpu
debugpy
distro
pymupdf
typing_inspect
py-cpuinfo
jiter
mypy_extensions
google-generativeai
hjson
patsy
websocket
py7zr
msgpack
python-Levenshtein
httpx-sse
fastavro
rouge-score
zstandard
vendi-score
torch
torchvision
numpy<2.0.0 # to avoid errors like "A module that was compiled using NumPy 1.x cannot be run in NumPy 2.0.0 as it may crash". related: https://github.com/huggingface/transformers/issues/31740. 
resiliparse
wget
uv
timeout-decorator
math-verify
pre-commit
rank-bm25